Moved applyImageControls to useCanvasControls
