+++ composables/useCanvasControls.ts
@@ -1,11 +1,23 @@
 // composables/useCanvasControls.ts
 
-import { FabricImage, Textbox } from 'fabric'
+import { FabricImage, Textbox, Control, controlsUtils } from 'fabric'
 import type { Canvas, FabricObject } from 'fabric'
 import { CircularTextbox } from '~/utils/canvasCircularTextbox'
-import { setTextboxTextRadius } from '~/utils/canvasUtils'
-import { useCanvasImage, attachSvgDataUrl } from '~/composables/useCanvasImage'
+import { setTextboxTextRadius, toggleObjectZOrder } from '~/utils/canvasUtils'
+import { attachSvgDataUrl } from '~/composables/useCanvasImage'
 import { useCanvasText } from '~/composables/useCanvasText'
+import {
+  createResizeControlRender,
+  createRotateControlRender,
+  createTrashControlRender,
+  createBringToFrontControlRender,
+} from '~/utils/canvasControlRenders'
+import { getResizeImage, getRotateImage, getTrashCanImage, getBringToFrontImage } from '@/utils/customImageIcons'
+
+
+interface Transform {
+  target?: unknown
+}
 
 /**
  * Canvas Controls Composable
@@ -19,8 +31,95 @@ import { useCanvasText } from '~/composables/useCanvasText'
  */
 
 export function useCanvasControls() {
-  const { applyImageControls } = useCanvasImage()
   const { applyTextboxControls } = useCanvasText()
+
+  /**
+   * Apply custom controls and appearance to a FabricImage.
+   * Called both when first adding and when restoring from JSON.
+   */
+  function applyImageControls(image: FabricImage): void {
+    // Clear default controls
+    image.controls = {}
+    // Configure the image
+    image.selectable = true
+    image.hasControls = true
+    image.hasBorders = true
+
+    // Add a blue dashed border for better visibility when selected
+    image.borderColor = 'blue'
+    image.borderScaleFactor = 1
+    image.borderDashArray = [5, 5]
+
+    // Disable caching to ensure controls are always rendered
+    image.objectCaching = false
+
+    // Add custom control for bring to front
+    image.controls.bringToFrontControl = new Control({
+      x: -0.5,
+      y: -0.5,
+      offsetX: -12,
+      offsetY: -12,
+      sizeX: 36,
+      sizeY: 36,
+      cursorStyle: 'pointer',
+      render: createBringToFrontControlRender(getBringToFrontImage()),
+      mouseUpHandler: (_eventData: unknown, transform: Transform): boolean => {
+        const target = transform?.target as FabricImage | undefined
+        if (target && target.canvas) {
+          toggleObjectZOrder(target, target.canvas)
+        }
+        return true
+      },
+    })
+
+    // Add custom control for delete
+    image.controls.deleteControl = new Control({
+      x: 0.5,
+      y: -0.5,
+      offsetX: 12,
+      offsetY: -12,
+      sizeX: 36,
+      sizeY: 36,
+      cursorStyle: 'pointer',
+      render: createTrashControlRender(getTrashCanImage()),
+      mouseUpHandler: (_eventData: unknown, transform: Transform): boolean => {
+        const target = transform?.target as FabricImage | undefined
+        if (target) {
+          const canvas = target.canvas
+          canvas?.remove(target)
+          canvas?.requestRenderAll()
+        }
+        return true
+      },
+    })
+    // Add custom control for rotation
+    image.controls.rotateControl = new Control({
+      x: 0,
+      y: -0.5,
+      offsetX: 0,
+      offsetY: -50,
+      sizeX: 36,
+      sizeY: 36,
+      cursorStyle: 'grab',
+      render: createRotateControlRender(getRotateImage()),
+      actionHandler: controlsUtils.rotationWithSnapping,
+      withConnection: true,
+    })
+
+    // Add custom control for resize
+    image.controls.resizeControl = new Control({
+      x: 0.5,
+      y: 0.5,
+      offsetX: 12,
+      offsetY: 12,
+      sizeX: 36,
+      sizeY: 36,
+      cursorStyle: 'nwse-resize',
+      render: createResizeControlRender(getResizeImage()),
+      withConnection: false,
+      actionHandler: controlsUtils.scalingEqually,
+    })
+  }
 
   function reapplyControlsToObject(obj: FabricObject): void {
     if (obj instanceof FabricImage) {
@@ -61,5 +160,5 @@ export function useCanvasControls() {
     }
   }
 
-  return { reapplyControls, reapplyControlsToObject, makeControlsReviver }
+  return { applyImageControls, reapplyControls, reapplyControlsToObject, makeControlsReviver }
 }
